4, 5, 6 tests

+++ Combination/CalculateCombination.cs
@@ -2,7 +2,7 @@
 {
     public class CalculateCombination
     {
-        private static int Factorial(int n)
+        private static long Factorial(int n)
         {
             if (n == 0 || n == 1)
             {
@@ -10,9 +10,9 @@
             }
             return n * Factorial(n - 1);
         }
-        public static int Calculate(int n, int k)
+        public static long Calculate(int n, int k)
         {
-            int result=Factorial(n)/(Factorial(n-k)*Factorial(k));
+            long result=Factorial(n)/(Factorial(n-k)*Factorial(k));
             return result;
         }
     }
